requirements.txt
langchain
langchain-community
faiss-cpu
sentence-transformers
huggingface-hub
python-dotenv
streamlit
pypdf
